+++ src/main/java/retry/domain/Cars.java
@@ -6,8 +6,6 @@ import java.util.Random;
 
 public class Cars {
 
-    private static final int PIVOT_NUMBER = 4;
-    private static final int RANDOM_BOUND_NUMBER = 9;
     private List<Car> cars;
 
     public Cars(List<Car> cars) {
@@ -18,16 +16,12 @@ public class Cars {
         return this.cars.contains(test);
     }
 
-    private int getRandomNumber() {
-        return new Random().nextInt(RANDOM_BOUND_NUMBER);
-    }
-
-    public List<Car> moveCarsPositionByInterface(Movable movable) {
+    public Cars moveCarsPositionByInterface(Movable movable) {
         List<Car> carsWithPosition = new ArrayList<>();
         for (Car car : this.cars) {
             carsWithPosition.add(car.move(movable.isMove()));
         }
-        return carsWithPosition;
+        return new Cars(carsWithPosition);
     }
 
 }
